Decoding a media file to PCM and playing it back via OpenSL ES.
Added audio_decoder.cpp file that is essentially a copy of the audio_recorder.cpp, but instead of microphone as data source, an URI is provided. This way it turns into a decoder to PCM. My goal is to modify the RAW PCM data and apply audio effects on it before playing it back. The decoder decodes data until the play buffer is full, then pauses and waits when the buffer is near empty, and then starts decoding again. I didn't want to store all decoded PCM data in memory, as it could be huge.

I could not make the decoder looping, nor I could change the play head position, so I had to re-create the decoder, in order to loop a media file. If someone knows how to restart the decoder, without recreating it, please share with me.

http://mobilepearls.com/labs/native-android-api/ndk/docs/opensles/
"Decode to PCM supports pause and initial seek. Volume control, effects, looping, and playback rate are not supported."